# src/main/resources/contract/api.yaml
openapi: 3.0.0
info:
  title: Proyecto DAW - Empresa de Tatuajes
  description: API para manejar el back-end de una empresa de tatuajes.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Servidor local
paths:
...
  /roles:
    get:
      summary: Obtener todos los roles
      description: "Solo los usuarios con rol ADMIN pueden obtener todos los roles."
      operationId: getRoles
      tags:
        - Roles
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                title: RoleResponse
                properties:
                  Roles:
                    type: array
                    description: Lista de roles
                    items:
                      $ref: "#/components/schemas/Role"

        '401':
          description: No autorizado (Unauthorized)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Acceso denegado (Forbidden)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /roles/{idRole}:
    put:
      summary: Actualizar un rol existente
      description: "Solo los usuarios con rol ADMIN pueden actualizar un rol."
      operationId: putRoleId
      tags:
        - Roles
      parameters:
        - in: path
          name: idRole
          required: true
          schema:
            type: string
          description: identificador del rol
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Role"
      responses:
        '200':
          description: Rol actualizado con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '400':
          description: Datos inválidos enviados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No autorizado (Unauthorized)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Acceso denegado (Forbidden)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Rol no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /Products:
    get:
      summary: Obtener todos los productos
      description: "Permite a cualquier usuario obtener la lista de productos disponibles."
      operationId: getProducts
      tags:
        - Products
      responses:
        "200":
          description: Lista de productos obtenida exitosamente
          content:
            application/json:
              schema:
                type: object
                title: ProductsResponse
                properties:
                  products:
                    type: array
                    description: Lista de productos disponibles
                    items:
                      $ref: "#/components/schemas/Products"
        "404":
          description: No se encontraron productos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Crear un nuevo producto
      description: "Permite a los administradores crear un nuevo producto proporcionando los datos requeridos."
      operationId: createProduct
      tags:
        - Products
      security:
        - bearerAuth: []
      requestBody:
        description: Datos del nuevo producto
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Products"
      responses:
        "201":
          description: Producto creado exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Products"
        "400":
          description: Solicitud incorrecta (Bad Request)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /Products/{idProduct}:
    get:
      summary: Obtener un producto por ID
      description: "Permite a cualquier usuario obtener los detalles de un producto específico mediante su ID único."
      operationId: getProductById
      tags:
        - Products
      parameters:
        - name: idProduct
          in: path
          required: true
          description: ID único del producto a obtener
          schema:
            type: integer
      responses:
        "200":
          description: Producto encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Products"
        "404":
          description: Producto no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Actualizar un producto por ID
      description: "Permite a los administradores actualizar los datos de un producto específico mediante su ID único."
      operationId: updateProduct
      tags:
        - Products
      security:
        - bearerAuth: []
      parameters:
        - name: idProduct
          in: path
          required: true
          description: ID único del producto a actualizar
          schema:
            type: integer
      requestBody:
        description: Nuevos datos del producto
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Products"
      responses:
        "200":
          description: Producto actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Products"
        "400":
          description: Solicitud incorrecta (Bad Request)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Producto no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Eliminar un producto por ID
      description: "Permite a los administradores eliminar un producto específico mediante su ID único."
      operationId: deleteProduct
      tags:
        - Products
      security:
        - bearerAuth: []
      parameters:
        - name: idProduct
          in: path
          required: true
          description: ID único del producto a eliminar
          schema:
            type: integer
      responses:
        "204":
          description: Producto eliminado exitosamente
        "404":
          description: Producto no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Role:
      type: object
      properties:
        idRole:
          type: integer
          format: int32
          description: Identificador único del rol
        roleName:
          type: string
          description: Nombre del rol

    # Objeto usuario actualizado (no tocar)
...
    Products:
      type: object
      properties:
        idProduct:
          type: integer
          format: int32
          description: Identificador único del producto
        idSubcategory:
          type: integer
          format: int32
          description: Identificador único de la subcategoría asociada
        name:
          type: string
          description: Nombre del producto
          maxLength: 100
        description:
          type: string
          description: Descripción detallada del producto
          nullable: true
        price:
          type: number
          format: double
          description: Precio del producto con exactamente dos decimales
        imageUrl:
          type: string
          description: URL o ruta de la imagen del producto
      nullable: true
      required:
        - name
        - price

    # Ordenes actualizadas (no tocar)
...
    ErrorResponse:
      type: object
      properties:
        errorCode:
          type: string
          description: Código del error
        errorMessage:
          type: string
          description: Descripción del error
security:
  - bearerAuth: []